service/requirements.txt
Flask==0.11
auth0-python==3.0.0
